Make test run and add wasmer error handing
Canon env, error processing, params fixes to run counter test

Make counter tests run

Make blockheight test work

Eliminate with-capacity

Fix (host) transact return value (wip)

Fix self snapshot test

Make 7 tests work

Make 11 tests work

Make all tests work

Make stack tests run

+++ src/ops/gas.rs
@@ -12,10 +12,10 @@ use crate::resolver::Env;
 pub struct Gas;
 
 impl Gas {
-    pub fn gas(env: &Env, gas_charged: u64) -> Result<(), VMError> {
+    pub fn gas(env: &Env, gas_charged: u32) -> Result<(), VMError> {
         let context: &mut CallContext = unsafe { &mut *(env.context.0 as *mut CallContext)};
         let meter = context.gas_meter_mut();
-        if meter.charge(gas_charged).is_out_of_gas() {
+        if meter.charge(gas_charged as u64).is_out_of_gas() {
             return Err(VMError::OutOfGas);
         }
         Ok(())
